+++ src/main/java/ba/unsa/etf/nbp24t1/entity/HallEntity.java
@@ -32,5 +32,5 @@ public class HallEntity {
     private Integer number;
 
     @Column(name = "cinema_id")
-    private Integer cinemaId;
+    private Long cinemaId;
 }
