+++ src/main/java/ba/unsa/etf/nbp24t1/entity/SeatEntity.java
@@ -36,5 +36,5 @@ public class SeatEntity {
     private String type;
 
     @Column(name = "hall_id")
-    private Integer hallId;
+    private Long hallId;
 }
